Refactor compiler.rs
